refactor(chat): streamline chat module and enhance guards
- Removed unused WebSocket guards and handlers to clean up the chat module.
- Replaced the deprecated WsJwtGuard with WsAuthGuard for improved authentication.
- Updated the ChatGateway to utilize the new guards and removed redundant imports.
- Introduced ChatPaymentGuard to enforce payment validation before joining chat rooms.
- Adjusted the ChatModule to reflect the changes in guards and services, enhancing maintainability.
