--- src/modules/chat/gateway/handlers/index.ts
@@ -1,4 +1,0 @@
-export { ConnectionHandler } from './connection.handler';
-export { MessageHandler } from './message.handler';
-export { RoomHandler } from './room.handler';
-export { TypingHandler } from './typing.handler';
